Tentatively add system to ensure that there is a single source of truth for JWTs (cookies in browser, memory outside browser for now)
We have run into some issues where, with situations involving a NextJS app and possibly some redirects, the in-memory `store.tokens` becomes decoupled from the tokens as stored in cookies. This causes things like Core JS reading tokens, finding no tokens (in memory), and as a result deleting the tokens from the cookies.

This commit revamps the token store to always rely on cookies as the source of truth in the browser. Tokens are read from and written to cookies, with a convenience layer in between so they can still be accessed as `store.tokens.accessToken` etc.

TODOs after this commit: update unit tests (they are expected to fail right now) and clean up the code a bit.